examples/LLM_Workflows/modular_llm_stack/requirements.txt
cohere
datasets
numpy
openai
pandas
pinecone-client
sentence-transformers
sf-hamilton
sf-hamilton[visualization]
weaviate-client
click
